requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
banks==2.2.0
beautifulsoup4==4.14.2
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
dataclasses-json==0.6.7
defusedxml==0.7.1
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
fastapi==0.121.2
filetype==1.2.0
frozenlist==1.8.0
fsspec==2025.10.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-genai==1.50.1
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
greenlet==3.2.4
griffe==1.15.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
Jinja2==3.1.6
jira==3.10.5
jiter==0.12.0
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
llama-cloud==0.1.35
llama-cloud-services==0.6.54
llama-index==0.14.8
llama-index-cli==0.5.3
llama-index-core==0.14.8
llama-index-embeddings-openai==0.5.1
llama-index-indices-managed-llama-cloud==0.9.4
llama-index-instrumentation==0.4.2
llama-index-llms-openai==0.6.9
llama-index-readers-file==0.5.4
llama-index-readers-llama-parse==0.5.1
llama-index-tools-jira==0.2.1
llama-index-workflows==2.11.2
llama-parse==0.6.54
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
mcp==1.21.1
mdurl==0.1.2
multidict==6.7.0
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.2
numpy==2.3.5
oauthlib==3.3.1
openai==2.8.0
packaging==25.0
pandas==2.2.3
pillow==12.0.0
platformdirs==4.5.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.5
pypdf==6.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
pywin32==311
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.2.0
rpds-py==0.28.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
sse-starlette==3.0.3
starlette==0.49.3
striprtf==0.0.26
tenacity==9.1.2
tiktoken==0.12.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.38.0
websockets==15.0.1
wrapt==1.17.3
yarl==1.22.0
python-docx
PyPDF2
